+++ CHANGELOG.md
@@ -104,6 +104,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - stored: add dedupable backend [PR #1663]
 - systemtests: fix plugin postgresql testrunner-* [PR #1947]
 - webui: fix handling deleted clients in restore browser [PR #1933]
+- console.cc: forbid @exec etc. as privileged user [PR #1950]
 
 ### Removed
 - plugins: remove old deprecated postgres plugin [PR #1606]
@@ -271,6 +272,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [PR #1936]: https://github.com/bareos/bareos/pull/1936
 [PR #1945]: https://github.com/bareos/bareos/pull/1945
 [PR #1947]: https://github.com/bareos/bareos/pull/1947
+[PR #1950]: https://github.com/bareos/bareos/pull/1950
 [PR #1956]: https://github.com/bareos/bareos/pull/1956
 [PR #1972]: https://github.com/bareos/bareos/pull/1972
 [unreleased]: https://github.com/bareos/bareos/tree/master